Views Service and domain classes refactoring

+++ src/main/java/org/apostolis/service/OperationsServiceImpl.java
@@ -13,6 +13,7 @@ import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.HashMap;
 
@@ -25,25 +26,35 @@ public class OperationsServiceImpl implements OperationsService{
     private final int max_comments_number;
 
     private final OperationsRepository operationsRepository;
+
+    private final RequestValidationService requestValidationService;
     private final TokenManager tokenManager;
 
     public OperationsServiceImpl(
-             OperationsRepository operationsRepository,
-             TokenManager tokenManager,
-             int free_post_size,
-             int premium_post_size,
-             int max_comments_number) {
+            OperationsRepository operationsRepository,
+            TokenManager tokenManager,
+            RequestValidationService requestValidationService,
+            int free_post_size,
+            int premium_post_size,
+            int max_comments_number) {
 
         this.operationsRepository = operationsRepository;
         this.tokenManager = tokenManager;
         this.free_post_size = free_post_size;
         this.premium_post_size = premium_post_size;
         this.max_comments_number = max_comments_number;
+        this.requestValidationService = requestValidationService;
     }
 
 
     @Override
     public void createPost(Post postToSave, String token) {
+
+        int authenticated_user_id = requestValidationService.extractUserId(token);
+        if (postToSave.getUser() != authenticated_user_id){
+            throw new UnauthorizedResponse("You are not allowed to make a post as another user");
+        }
+
         Role authlevel = tokenManager.extractRole(token);
         int post_size = postToSave.getText().length();
         switch (authlevel){
@@ -62,11 +73,17 @@ public class OperationsServiceImpl implements OperationsService{
             default:
                 break;
         }
+        //postToSave.setCreatedAt(LocalDateTime.now(AppConfig.clock));
         operationsRepository.savePost(postToSave);
     }
 
     @Override
     public void createComment(Comment commentToSave, String token) {
+
+        int authenticated_user_id = requestValidationService.extractUserId(token);
+        if (commentToSave.getUser() != authenticated_user_id){
+            throw new UnauthorizedResponse("You are not allowed to make a comment as another user");
+        }
 
         if(tokenManager.extractRole(token).equals(Role.FREE)){
             int comments_count = operationsRepository.getCountOfUserCommentsUnderThisPost(
@@ -77,10 +94,15 @@ public class OperationsServiceImpl implements OperationsService{
                         "\nYou reached the maximum number of comments for this post.");
             }
         }
+        //commentToSave.setCreatedAt(LocalDateTime.now(AppConfig.clock));
         operationsRepository.saveComment(commentToSave);
     }
     @Override
-    public void follow(int follower, int to_follow) {
+    public void follow(int follower, int to_follow, String token) {
+        int authenticated_user_id = requestValidationService.extractUserId(token);
+        if (follower != authenticated_user_id){
+            throw new UnauthorizedResponse("Your request id does not match with your authentication id");
+        }
         if (follower != to_follow){
             operationsRepository.saveFollow(follower, to_follow);
         }else{
@@ -89,13 +111,26 @@ public class OperationsServiceImpl implements OperationsService{
     }
 
     @Override
-    public void unfollow(int follower, int to_unfollow) {
+    public void unfollow(int follower, int to_unfollow, String token) {
+        int authenticated_user_id = requestValidationService.extractUserId(token);
+        if (follower != authenticated_user_id){
+            throw new UnauthorizedResponse("Your request id does not match with your authentication id");
+        }
         operationsRepository.deleteFollow(follower, to_unfollow);
     }
 
 
     @Override
-    public String createUrlForPostAndComments(int user, int post) {
+    public String createUrlForPostAndComments(int user, int post, String token) {
+        int authenticated_user_id = requestValidationService.extractUserId(token);
+        if(user != authenticated_user_id){
+            throw new UnauthorizedResponse("Your request id does not match with your authentication id");
+        }
+        ArrayList<Integer> user_post_ids = requestValidationService.getUsersPostIds(user);
+        if(!user_post_ids.contains(post)){
+            throw new UnauthorizedResponse("You cannot create shareable link for a post of another user");
+        }
+
         try{
             // register the link to prevent data leaks via url manipulation
             operationsRepository.registerLink(user, post);
@@ -128,5 +163,11 @@ public class OperationsServiceImpl implements OperationsService{
         }catch(UnsupportedEncodingException e){
             throw new RuntimeException(e.getMessage());
         }
+    }
+
+
+    @Override
+    public String getUsername(int user_id){
+        return requestValidationService.extractUsername(user_id);
     }
 }
